refactor: extract SourceKind type and useSourceKinds hook BED-8743
