+++ packages/javascript/js-client-library/src/client.ts
@@ -107,6 +107,7 @@ import {
     RotateWebhookSecretResponse,
     SavedQuery,
     SavedQueryPermissionsResponse,
+    SourceKindsResponse,
     StartFileIngestResponse,
     UnifiedFindingResponse,
     UpdateConfigurationResponse,
@@ -256,10 +257,7 @@ class BHEAPIClient {
     getKinds = (options?: RequestOptions) => this.baseClient.get<GraphKindsResponse>('/api/v2/graphs/kinds', options);
 
     getSourceKinds = (options?: RequestOptions) =>
-        this.baseClient.get<BasicResponse<{ kinds: { id: number; name: string }[] }>>(
-            '/api/v2/graphs/source-kinds',
-            options
-        );
+        this.baseClient.get<SourceKindsResponse>('/api/v2/graphs/source-kinds', options);
 
     clearDatabase = (payload: ClearDatabaseRequest, options?: RequestOptions) => {
         return this.baseClient.post('/api/v2/clear-database', payload, options);
